requirements.txt
anyio==3.7.1
asyncpg==0.27.0
click==8.1.3
databases==0.7.0
fastapi==0.99.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
pydantic==1.10.11
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
